feat(v1): Environment consumes the Agent — single-agent as the default program
The Environment is now framed as a program of agent interactions over a
taskset, running the simplest one: every episode's rollouts are built
through an internal Agent, so an eval run and a hand-written agent
program resolve placement, timeouts, and pairing through one path.

- verifiers/v1/resolve.py (new): per-run resolution — TimeoutConfig,
  resolve_runtime_config, validate_pairing, cap_remote_harness_timeout —
  moved below both consumers (breaks the agent <- env import cycle).
- Agent.run splits into rollout() (resolve-only, returns the Rollout;
  accepts the eval's serving resources) + run() (rollout + await).
- Rollout.run stamps trace.info["agent"] itself: every trace — eval or
  program — carries provenance now.
- Environment drops its mirrored resolution (runtime_for, per-stage
  timeout fields); episode() goes through Agent.rollout.
